requirements.txt
reportlab
qrcode
pyyaml
